% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ExPanD_config_worldbank}
\alias{ExPanD_config_worldbank}
\title{Default Configuration to Use with ExPanD and the \code{worldbank} Data Set}
\format{An object of class \code{"list"}.}
\usage{
data(ExPanD_config_worldbank)
}
\description{
List to use as a \code{list_config} parameter when starting \link{ExPanD}.
}
\examples{
data(worldbank)
data(worldbank_data_def)
data(worldbank_var_def)
data(ExPanD_config_worldbank)
\dontrun{
  ExPanD(worldbank, df_def = worldbank_data_def,
    var_def = worldbank_var_def, config_list = ExPanD_config_worldbank)
}
}
\keyword{ExPanD,}
\keyword{config}
