% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExPanDaR.R
\docType{package}
\name{ExPanDaR}
\alias{ExPanDaR}
\title{ExPanDaR: Explore Panel Data with R}
\description{
ExPanDaR provides the code base for the ExPanD web app.
ExPanD is a shiny based app supporting interactive
exploratory data analysis.
}
\details{
ExPanDaR has two main goals:
\itemize{
  \item{} Enable users to assess the robustness of
    empirical evidence without providing them with
    access to the underlying data.
  \item{}  Provide a toolbox for researchers to explore
    panel data on the fly.
}
To learn more about ExPanDaR, start with the vignettes:
\code{browseVignettes(package = "ExPanDaR")}
}
