% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepapre_correlation_graph.R
\name{prepare_correlation_graph}
\alias{prepare_correlation_graph}
\title{Prepares a Correlation Graph}
\usage{
prepare_correlation_graph(df)
}
\arguments{
\item{df}{Data frame containing at least two variables that are either numeric
or logical and at least five observations.}
}
\value{
The graph as prepared by corrplot
}
\description{
Reads a data frame and presents Pearson correlations above the diagonal
and Spearman correlations using a fany graph prepared by the package \code{corrplot}.
}
\examples{
prepare_correlation_graph(mtcars)
}
