% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_ext_obs_table.R
\name{prepare_ext_obs_table}
\alias{prepare_ext_obs_table}
\title{Prepares a Table Displaying Extreme Observations}
\usage{
prepare_ext_obs_table(df, n = 5, ...)
}
\arguments{
\item{df}{Data frame containing at least one variable that is numeric as variable of
interest. Most likely it will also contain the variables identifying an observation.}

\item{n}{The number of top/bottom observations that you want to report.}

\item{...}{Additional parameters that are passed to \code{knitr::kable()}.}
}
\value{
A list containing two items:
\itemize{
 \item{"df"}{A data frame containg the top/bottom n observations}
 \item{"kable_ret"}{The return value provided by \code{knitr::kable()} containing the formatted table}
}
}
\description{
Reads a data frame containing the variable of interest as the last variable,
  sorts it by it and displays the top and bottom n variables.
}
\details{
The default parameters for calling \code{knitr::kable},
  are \code{format = "html", digits = 3, format.args = list(big.mark = ','), row.names = FALSE}.
}
\examples{
t <- prepare_ext_obs_table(data.frame(name = rownames(mtcars), hp = mtcars$hp))
t$kable_ret
}
