% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepapre_missing_values_graph.R
\name{prepare_missing_values_graph}
\alias{prepare_missing_values_graph}
\title{Prepares a Graph Displaying Missing Values in Panel Data}
\usage{
prepare_missing_values_graph(df, period)
}
\arguments{
\item{df}{Data frame containing the data. Only variables containing
numerical or logical data will be used.}

\item{period}{A string containing the name of the factor indicating the time dimension.}
}
\value{
A ggplot2 graph
}
\description{
Displays a heatmap of missing value frequency across the panel
}
\details{
This was inspireed by a blog post of Rense Nieuenhuis. Thanks!
  http://www.rensenieuwenhuis.nl/r-sessions-30-visualizing-missing-values/
}
\examples{
df <- data.frame(year = as.ordered(floor(time(EuStockMarkets))), EuStockMarkets)
period <- "year"
prepare_missing_values_graph(df, period="year")
}
