% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_scatter_plot.R
\name{prepare_scatter_plot}
\alias{prepare_scatter_plot}
\title{Prepares a Scatter Plot}
\usage{
prepare_scatter_plot(df, x, y, color = NULL, size = NULL, loess = 0)
}
\arguments{
\item{df}{Data frame containing at most four numerical variables}

\item{x}{a string containing the column name of the x variable}

\item{y}{a string containing the column name of the x variable}

\item{color}{a string containing the column name of the x variable
(can be numerical or a factor)}

\item{size}{a string containing the column name of the x variable}

\item{loess}{a numerical scalar string
\describe{
 \item{"0"}{No loess curve}
 \item{"1"}{loess curve with equal weights}
 \item{"2"}{loess curve with weights based on \code{size} variable}
}}
}
\value{
the plot as returned by ggplot
}
\description{
Reads a data frame containing at most four numerical variables and prepares a scatter plot.
}
