% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_trend_graph.R
\name{prepare_trend_graph}
\alias{prepare_trend_graph}
\title{Prepares a Trend Graph}
\usage{
prepare_trend_graph(df, xvar)
}
\arguments{
\item{df}{Data frame containing the ordered factor and a set of numerical variables
to be plotted}

\item{xvar}{a string countaining the column name of the ordered factor}
}
\value{
A list containing two items:
\describe{
 \item{"df"}{A data frame containg the plotted means and standard errors}
 \item{"plot}{The plot as returned by ggplot}
}
}
\description{
Reads a data frame and line plots all variables (which need to be numeric)
 by an ordered factor.
}
\examples{
df <- data.frame(year = floor(time(EuStockMarkets)), EuStockMarkets)
graph <- prepare_trend_graph(df, "year")
graph$plot
}
