% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{russell_3000}
\alias{russell_3000}
\title{Accounting and Annual Stock Return Data for a Sample of Russell 3000 Firms (2013-2016)}
\format{An object of class \code{"data.frame"}.}
\source{
Has been collected using the \code{\link[tidyquant]{tq_get}} function family in Summer 2017.
 The data has been collected for the sole purpose to showcase the functions of the \code{ExPanDaR} package.
 Use in scientific studies is strongly discouraged.
}
\usage{
data(russell_3000)
}
\description{
Data collected from Google Finance and Yahoo finance using the package \code{tidyquant}.
}
\examples{
data(russell_3000)
prepare_missing_values_graph(russell_3000, period = "period")

}
\keyword{accounting}
\keyword{capital}
\keyword{data}
\keyword{data,}
\keyword{datasets,}
\keyword{market}
