% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{russell_3000_data_def}
\alias{russell_3000_data_def}
\title{Data Definitions for \code{russell_3000} Data Set}
\format{An object of class \code{"data.frame"}.}
\usage{
data(russell_3000_data_def)
}
\description{
A data frame containing variable definitions for the \code{russell_3000} data set.
The data definitions can be passed to \link{ExPanD} via the
\code{df_def} parameter.
}
\details{
Data definitions are provided by the package maintainer and are somewhat superficial
to make them both, short and informative. User discretion is advised when using this
data outside of its didactic purpose.
}
\examples{
data(russell_3000)
data(russell_3000_data_def)
data(ExPanD_config_russell_3000)
\dontrun{
  ExPanD(russell_3000, df_def = russell_3000_data_def, config_list = ExPanD_config_russell_3000)
}
}
\keyword{datasets}
