% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treat_outliers.R
\name{treat_outliers}
\alias{treat_outliers}
\title{A function to treat outliers in numerical data}
\usage{
treat_outliers(x, percentile = 0.01, truncate = FALSE, byvec = NULL)
}
\arguments{
\item{x}{A numeric vector or matrix.}

\item{percentile}{A numeric scalar.
  The percentile below which observations
  are considered to be outliers. Is treated symmetrical so that
\code{c(percentile, 1-percentile)} are used as boundaries.
  Defaults to 0.01.}

\item{truncate}{A logical scalar. If TRUE then data are truncated
(i.e., set to NA if out of bounds). Defaults to FALSE.}

\item{byvec}{NULL or a factor vector containing groups
by which the outlier treatment is applied. Defaults to NULL.
If provided, it needs to be such so that \code{length(byvec) == nrows(as.matrix(x))}.}
}
\value{
A numeric vector or matrix containing the outlier-treated \code{x}.
}
\description{
Treats numerical outliers either by winsorizing or by truncating.
}
\examples{
treat_outliers(seq(1:100), 0.05)
treat_outliers(seq(1:100), truncate = TRUE, 0.05)
summary(treat_outliers(data.frame(a = seq(1:100), b = rnorm(100)), percentile = 0.05))
hist(treat_outliers(rnorm(1000)), breaks=100)
}
