% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{worldbank_var_def}
\alias{worldbank_var_def}
\title{Variable Definitions to Construct an Analysis Sample Based on the \code{worldbank} Data Set}
\format{An object of class \code{"data.frame"}.}
\usage{
data(worldbank_var_def)
}
\description{
A data frame containing variable definitions that can be passed to \code{\link{ExPanD}}
via the var_def parameter.
}
\examples{
data(worldbank)
data(worldbank_data_def)
data(worldbank_var_def)
data(ExPanD_config_worldbank)
\dontrun{
  ExPanD(worldbank, df_def = worldbank_data_def,
    var_def = worldbank_var_def, config_list = ExPanD_config_worldbank)
}
}
\keyword{datasets}
