% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExPanD.R
\name{ExPanD}
\alias{ExPanD}
\title{Explore Panal Data (ExPanD)}
\usage{
ExPanD(df, cs_id, ts_id, ...)
}
\arguments{
\item{df}{A data that containing the panel data that you want to explore}

\item{cs_id}{A character vector containing the names of the variables that
identify the cross-section in your data. The according variables should be factors.}

\item{ts_id}{A character scalar identify the name of the variables that
identifies the time series in your data. The according variables needs to be anordered vector.}
}
\description{
A shiny based app that uses ExPanDaR functionality for interactive
 panel data exploration
}
\details{
Use the \code{\link[Hmisc]{label}} command to add label defining the variables to your data frame if you want
  to have nice floatting variable descriptions in the descriptive table.
}
\examples{
data(russell_3000)
ExPanD(russell_3000, c("coid", "coname"), "period")

}
