% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ExPanD_config_russell_3000}
\alias{ExPanD_config_russell_3000}
\title{Default Configuration to use with ExPanD and the Russell 3000 Data Set}
\format{An object of class \code{"list"}.}
\usage{
data(ExPanD_config_russell_3000)
}
\description{
List to use as a \code{list_config} parameter when starting \link{ExPanD}.
}
\examples{
data(russell_3000)
data(russell_3000_data_def)
data(ExPanD_config_russell_3000)
\dontrun{
  ExPanD(russell_3000, df_def = russell_3000_data_def, config_list = ExPanD_config_russell_3000)
}
}
\keyword{datasets}
