% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepapre_missing_values_graph.R
\name{prepare_missing_values_graph}
\alias{prepare_missing_values_graph}
\title{Prepares a Graph Displaying Missing Values in Panel Data}
\usage{
prepare_missing_values_graph(df, ts_id, no_factors = FALSE)
}
\arguments{
\item{df}{Data frame containing the data. Only variables containing
numerical or logical data will be used.}

\item{ts_id}{A string containing the name of the variable indicating the time dimension.
Needs to be coercible into an ordered factor.}

\item{no_factors}{A logical variable indicating whether you want to limit the plot to
logical and numerical variables. Defaults to FALSE.}
}
\value{
A ggplot2 graph
}
\description{
Displays a heatmap of missing value frequency across the panel
}
\details{
This was inspireed by a blog post of Rense Nieuenhuis. Thanks!
  http://www.rensenieuwenhuis.nl/r-sessions-30-visualizing-missing-values/
}
\examples{
df <- data.frame(year = as.ordered(floor(time(EuStockMarkets))), EuStockMarkets)
ts_id <- "year"
prepare_missing_values_graph(df, ts_id="year")
}
