% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_quantile_trend_graph.R
\name{prepare_quantile_trend_graph}
\alias{prepare_quantile_trend_graph}
\title{Prepares a Quantile Trend Graph}
\usage{
prepare_quantile_trend_graph(df, xvar, quantiles = c(0.05, 0.25, 0.5, 0.75,
  0.95))
}
\arguments{
\item{df}{Data frame containing the ordered factor and the numerical variable
to be plotted}

\item{xvar}{a string containing the column name of the ordered factor}

\item{quantiles}{a numerical vector containing the quantiles that are to be plotted}
}
\value{
A list containing two items:
\describe{
 \item{"df"}{A data frame containg the plotted quantiles}
 \item{"plot"}{The plot as returned by ggplot}
}
}
\description{
Reads a data frame containing two variables (an ordered factor and a numeric variable)
and plots the quantiles of the numeric variables
by the ordered factor.
}
\examples{
df <- data.frame(year = floor(stats::time(datasets::EuStockMarkets)),
                 DAX = datasets::EuStockMarkets[,"DAX"])
graph <- prepare_quantile_trend_graph(df, "year", c(0.05, 0.25, 0.5, 0.75, 0.95))
graph$plot
}
