% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepapre_regression_table.R
\name{prepare_regression_table}
\alias{prepare_regression_table}
\title{Prepares a Regression Table}
\usage{
prepare_regression_table(df, dvs, idvs, feffects = "", clusters = "",
  byvar = "", format = "html")
}
\arguments{
\item{df}{Data frame containing the data to estimate the models on.}

\item{dvs}{A character vector containing the dependent variable(s).}

\item{idvs}{A character vector or a a list of character vectors containing the independent variables.}

\item{feffects}{A character vector or a a list of character vectors containing the fixed effects.}

\item{clusters}{A character vector or a a list of character vectors containing the cluster variables.}

\item{byvar}{A factorial variable to estimate the model on (only possible if only one model is being estimated).}

\item{format}{A character scalar that is passed on \link[stargazer]{stargazer} as \code{type} to determine the presentation
format (e.g., "html", "text", or "latex").}
}
\value{
A list contining two items
\describe{
 \item{"models"}{A list containg the model results}
 \item{"table"}{The output of \link[stargazer]{stargazer} containing the table}
}
}
\description{
Builds a regression table based on a set of user-specifed models or a single model and a partitioning variable.
}
\details{
Depending on whether the dependent variable is numeric or a factor with two levels, the models are estimated
  using \code{\link[lfe]{felm}} or \link[stats]{glm} (with \code{family = binomial(link="logit")}).
  Fixed effects and clustered standard errors are only supported with continous dependent variables.
}
\examples{
df <- data.frame(year = as.factor(floor(stats::time(datasets::EuStockMarkets))),
                 datasets::EuStockMarkets)
dvs = list("DAX", "SMI", "CAC", "FTSE")
idvs = list(c("SMI", "CAC", "FTSE"),
            c("DAX", "CAC", "FTSE"),
            c("SMI", "DAX", "FTSE"),
            c("SMI", "CAC", "DAX"))
feffects = list("year", "year", "year", "year")
clusters = list("year", "year", "year", "year")
t <- prepare_regression_table(df, dvs, idvs, feffects, clusters)
t$table
t <- prepare_regression_table(df, "DAX", c("SMI", "CAC", "FTSE"), byvar="year")
print(t$table)
}
